add tablets page and accessories page
