+++ src/components/PageContent/PageContent.scss
@@ -1,6 +1,6 @@
 @import "../../styles/utils/variables";
 
-.PhonesPage {
+.PageContent {
   padding: 24px 0 80px;
 
   &__breadcrumbs-wrapper {
