+++ src/pages/TabletsPage/TabletsPage.tsx
@@ -1,5 +1,12 @@
+import { PageContent } from '../../components/PageContent';
+import { MenuItems } from '../../types/MenuItems';
+import { getTablets } from '../../helpers/api';
+
 export const TabletsPage = () => {
   return (
-    <h1>Tablets</h1>
+    <PageContent
+      title={MenuItems.Tablets}
+      getProducts={getTablets}
+    />
   );
 };
